[Update] V1 API - Moved v1 routes to API/v1
