+++ config/app.php
@@ -40,7 +40,7 @@ return [
     | or any other location as required by the application or its packages.
     */
 
-    'version' => '1.13.6-alpha.1',
+    'version' => '1.14.0-alpha.1',
 
     /*
     |--------------------------------------------------------------------------
