+++ routes/API/v1/Anime.php
