+++ routes/API/v1/Genres.php
