+++ routes/API/v1/Legal.php
